requirements.txt
accelerate==1.8.1
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
diffusers==0.34.0
Django==5.1
filelock==3.18.0
Flask==3.1.1
fpdf==1.7.2
fsspec==2025.5.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
httplib2==0.22.0
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.5
numpy==2.1.0
oauth2client==4.1.3
oauthlib==3.3.0
opencv-python==4.10.0.84
packaging==25.0
pillow==11.3.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.1
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.4.0
Werkzeug==3.1.3
win10toast==0.9
zipp==3.23.0
